requirements.txt
asgiref==3.7.2
asttokens==2.4.0
backcall==0.2.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
comm==0.1.4
cryptography==41.0.4
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.8.0rc2
Django==4.2.5
django-cors-headers==4.2.0
django-mssql-backend-azure==2.9.2
django-pyodbc-azure==2.1.0.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
et-xmlfile==1.1.0
executing==2.0.0
idna==3.4
ipykernel==6.25.2
ipython==8.16.1
jedi==0.19.1
jupyter_client==8.4.0
jupyter_core==5.4.0
matplotlib-inline==0.1.6
msal==1.24.0
mssql-django==1.3
nest-asyncio==1.5.8
numpy==1.26.0
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.2
pandas==2.1.1
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.6
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
PyMySQL==1.1.0
pyodbc==5.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
pywin32
# --pywin32==306
pyzmq==25.1.1
requests==2.31.0
requests-oauthlib==1.3.1
setuptools==68.2.2
six==1.16.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
stack-data==0.6.3
tornado==6.3.3
traitlets==5.11.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
wcwidth==0.2.8
wheel==0.41.2
whitenoise==6.6.0
